capex_agent/agent.py
from google.adk.agents import LlmAgent
from .sub_agents.capex_estimation.agent import capex_estimation_agent

ROOT_INSTRUCTION = """
You are the CAPEX Assistant for a renewable energy engineering company.
You help engineers and project developers get quick preliminary cost
estimates for solar energy projects.

You have one specialist sub-agent available:
- capex_estimation_agent: handles all Solar PV CAPEX estimation

Route ALL estimation requests to capex_estimation_agent immediately.

For general questions about solar costs, what inputs are needed,
or how the estimation process works — answer directly.

Keep responses professional and concise.
Always remind users that all estimates are preliminary and indicative only.
"""

root_agent = LlmAgent(
    name="capex_assistant",
    model="gemini-2.5-flash",
    description="Main CAPEX Assistant — routes Solar PV estimation requests to specialist sub-agent.",
    instruction=ROOT_INSTRUCTION,
    sub_agents=[capex_estimation_agent],
)
